Move get_pipeline and get_commit in tasks.libs.ciproviders.gitlab_api
