+++ tasks/libs/notify/alerts.py
@@ -8,7 +8,7 @@ from datetime import datetime, timezone
 from invoke.context import Context
 from invoke.exceptions import UnexpectedExit
 
-from tasks.libs.ciproviders.gitlab_api import BASE_URL
+from tasks.libs.ciproviders.gitlab_api import BASE_URL, get_pipeline
 from tasks.libs.common.datadog_api import create_count, send_metrics
 from tasks.libs.notify.utils import (
     AWS_S3_CP_CMD,
@@ -17,7 +17,7 @@ from tasks.libs.notify.utils import (
     PROJECT_TITLE,
     get_ci_visibility_job_url,
 )
-from tasks.libs.pipeline.data import get_failed_jobs, get_pipeline
+from tasks.libs.pipeline.data import get_failed_jobs
 from tasks.libs.pipeline.notifications import (
     get_pr_from_commit,
     send_slack_message,
